Finished level design
still need to make icon interaction for generator shield, add some candles with fancy lights for atmosphere, add interactions for 3 holes in damaged  walls. thats it about my idea!

Which bugs I couldn't fix: poster interaction ui works a bit ugly so I hope Connor will manage it, and also Vent seems not to be working properly in terms of the suitcase which keeps fall in textures by half and the padlock's interaction is available without opening the vent which should not be like that!