Created Basic Menu
I duplicated the chamber level and created a scene for menu there. It is functioning but not that beautiful for now. exit works when we build the game. start the game also needs the building as it will load the scene with index 1 which is going to be the chamber